refactor: clean scripts and dependencies
