+++ package.json
@@ -71,10 +71,9 @@
     "babel-preset-stage-3": "6.24.1",
     "babel-register": "^6.26.0",
     "babel-runtime": "6.26.0",
-    "codacy-coverage": "2.0.3",
     "codecov": "2.2.0",
     "coveralls": "2.13.1",
-    "cross-env": "^5.1.1",
+    "cross-env": "5.1.1",
     "css-loader": "0.28.7",
     "element-react": "1.4.3",
     "element-theme-default": "1.4.12",
@@ -140,9 +139,8 @@
     "pretest": "npm run code:build",
     "test": "cross-env jest",
     "pre:ci": "npm run build:webui",
-    "test:only": "mocha ./test/functional ./test/unit",
     "coverage:publish": "codecov",
-    "lint": "eslint .",
+    "lint": "npm run flow && eslint .",
     "lint:css": "stylelint 'src/**/*.scss' --syntax scss",
     "dev:start": "cross-env BABEL_ENV=registry babel-node src/lib/cli",
     "code:build": "cross-env BABEL_ENV=registry babel src/ --out-dir build/ --ignore src/webui/ --copy-files",
